Split static classes into their own files under a top package and update style
